Replace auth sub-app components with middleware
