+++ packages/shopify-app-express/src/auth/auth-callback.ts
@@ -7,25 +7,29 @@ import {
   Shopify,
 } from '@shopify/shopify-api';
 
-import {AppConfigInterface, AfterAuthCallback} from '../config-types';
-import {redirectToHost} from '../redirect-to-host';
+import {AppConfigInterface} from '../config-types';
+import {redirectToAuth} from '../redirect-to-auth';
 
 import {AuthCallbackParams} from './types';
-import {authBegin} from './auth-begin';
 
 export async function authCallback({
   req,
   res,
   api,
   config,
-  afterAuth,
 }: AuthCallbackParams) {
+  await config.logger.info('Initiating auth callback request');
+
   try {
     const callbackResponse = await api.auth.callback({
       isOnline: config.useOnlineTokens,
       rawRequest: req,
       rawResponse: res,
     });
+    res.locals.shopify = {
+      ...res.locals.shopify,
+      session: callbackResponse.session,
+    };
 
     await config.logger.debug('Callback is valid, storing session', {
       shop: callbackResponse.session.shop,
@@ -37,14 +41,7 @@ export async function authCallback({
       shop: callbackResponse.session.shop,
     });
 
-    await afterAuthActions(
-      req,
-      res,
-      config,
-      api,
-      callbackResponse.session,
-      afterAuth,
-    );
+    await afterAuthActions(req, res, config, api, callbackResponse.session);
   } catch (error) {
     await config.logger.error(`Failed to complete OAuth with error: ${error}`);
 
@@ -53,31 +50,13 @@ export async function authCallback({
 }
 
 async function afterAuthActions(
-  req: Request,
-  res: Response,
+  _req: Request,
+  _res: Response,
   config: AppConfigInterface,
   api: Shopify,
   session: Session,
-  afterAuth?: AfterAuthCallback,
 ) {
   await registerWebhooks(config, api, session);
-
-  if (afterAuth) {
-    await config.logger.debug('Calling afterAuth callback', {
-      shop: session.shop,
-    });
-
-    await afterAuth({
-      req,
-      res,
-      session,
-    });
-  }
-
-  // We redirect to the host-based app URL ONLY if the afterAuth callback didn't send a response already
-  if (!res.headersSent) {
-    await redirectToHost({req, res, api, config, session});
-  }
 }
 
 async function registerWebhooks(
@@ -129,7 +108,7 @@ async function handleCallbackError(
       res.send(error.message);
       break;
     case error instanceof CookieNotFound:
-      await authBegin({req, res, api, config});
+      await redirectToAuth({req, res, api, config});
       break;
     default:
       res.status(500);
