+++ packages/shopify-app-express/src/middlewares/index.ts
@@ -4,10 +4,12 @@ import {
   createEnsureInstalled,
 } from './ensure-installed-on-shop';
 import {createCspHeaders} from './csp-headers';
+import {redirectToShopifyOrAppRoot} from './redirect-to-shopify-or-app-root';
 
 export {
   createValidateAuthenticatedSession,
   createDeleteAppInstallationHandler,
   createCspHeaders,
   createEnsureInstalled,
+  redirectToShopifyOrAppRoot,
 };
