+++ packages/shopify-app-express/src/middlewares/types.ts
@@ -3,3 +3,4 @@ import {RequestHandler} from 'express';
 export type ValidateAuthenticatedSessionMiddleware = () => RequestHandler;
 export type EnsureInstalledMiddleware = () => RequestHandler;
 export type CspHeadersMiddleware = () => RequestHandler;
+export type RedirectToShopifyOrAppRootMiddleware = () => RequestHandler;
