+++ packages/shopify-app-express/src/shopify-app/types.ts
@@ -1,10 +1,8 @@
 import {Express} from 'express';
 
-import {AfterAuthCallback} from '../config-types';
 import {WebhookHandlersParam} from '../webhooks/types';
 
 export interface AppMiddlewareParams {
-  afterAuth?: AfterAuthCallback;
   webhookHandlers?: WebhookHandlersParam;
 }
